pyproject.toml
[project]
name = "borrowed-book-reminder"
version = "0.1.0"
description = "Synchronisation Médiathèque Villeurbanne → Nextcloud"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "caldav>=3.1.0",
]

[dependency-groups]
dev = [
    "ipython>=9.12.0",
]
